% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgwr.R
\name{coef.hgwrm}
\alias{coef.hgwrm}
\title{Get estimated coefficients.}
\usage{
\method{coef}{hgwrm}(object, ...)
}
\arguments{
\item{object}{An \code{hgwrm} object returned by \code{\link[=hgwr]{hgwr()}}.}

\item{\dots}{Parameter received from other functions.}
}
\value{
A \code{DataFrame} object consists of all estimated coefficients.
}
\description{
Get estimated coefficients.
}
\seealso{
\code{\link[=hgwr]{hgwr()}}, \code{\link[=summary.hgwrm]{summary.hgwrm()}}, \code{\link[=fitted.hgwrm]{fitted.hgwrm()}} and \code{\link[=residuals.hgwrm]{residuals.hgwrm()}}.
}
