% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgwr.R
\name{fitted.hgwrm}
\alias{fitted.hgwrm}
\title{Get fitted response.}
\usage{
\method{fitted}{hgwrm}(object, ...)
}
\arguments{
\item{object}{An \code{hgwrm} object returned by \code{\link[=hgwr]{hgwr()}}.}

\item{...}{Parameter received from other functions.}
}
\value{
A vector consists of fitted response values.
}
\description{
Get fitted response.
}
\seealso{
\code{\link[=hgwr]{hgwr()}}, \code{\link[=summary.hgwrm]{summary.hgwrm()}}, \code{\link[=coef.hgwrm]{coef.hgwrm()}} and \code{\link[=residuals.hgwrm]{residuals.hgwrm()}}.
}
