% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hetero.R, R/hgwr.R
\name{spatial_hetero_test.hgwrm}
\alias{spatial_hetero_test.hgwrm}
\alias{hgwr}
\alias{hgwr.sf}
\alias{hgwr.data.frame}
\alias{hgwr_fit}
\title{Hierarchical and Geographically Weighted Regression}
\usage{
\method{spatial_hetero_test}{hgwrm}(
  x,
  round = 99,
  statistic = stat_glsw,
  parallel = FALSE,
  verbose = 0,
  ...
)

hgwr(
  formula,
  data,
  ...,
  bw = "CV",
  kernel = c("gaussian", "bisquared"),
  eps_iter = 1e-06,
  max_iters = 1e+06,
  max_retries = 1e+06,
  global_options = ml_options(),
  f_test = FALSE,
  verbose = 0
)

\method{hgwr}{sf}(
  formula,
  data,
  ...,
  bw = "CV",
  kernel = c("gaussian", "bisquared"),
  eps_iter = 1e-06,
  max_iters = 1e+06,
  max_retries = 1e+06,
  global_options = ml_options(),
  f_test = FALSE,
  verbose = 0
)

\method{hgwr}{data.frame}(
  formula,
  data,
  ...,
  coords,
  bw = "CV",
  kernel = c("gaussian", "bisquared"),
  eps_iter = 1e-06,
  max_iters = 1e+06,
  max_retries = 1e+06,
  global_options = ml_options(),
  f_test = FALSE,
  verbose = 0
)

hgwr_fit(
  formula,
  data,
  coords,
  bw = c("CV", "AIC"),
  kernel = c("gaussian", "bisquared"),
  eps_iter = 1e-06,
  max_iters = 1e+06,
  max_retries = 1e+06,
  global_options = ml_options(),
  f_test = FALSE,
  verbose = 0
)
}
\arguments{
\item{x}{An \code{hgwrm} object}

\item{round}{The number of times to sampling from model.}

\item{statistic}{A function used to calculate the statistics
on the original data and bootstrapped data.
Default to the variance of standardlised GLSW estimates.}

\item{parallel}{If TRUE, use \code{furrr} package to parallel.}

\item{verbose}{An integer value. Determine the log level.
Possible values are:
\describe{
\item{0}{no log is printed.}
\item{1}{only logs in back-fitting are printed.}
\item{2}{all logs are printed.}
}}

\item{\dots}{Further arguments for the specified type of \code{data}.}

\item{formula}{A formula.
Its structure is similar to \code{\link[lme4]{lmer}} function
in \strong{lme4} package.
Models can be specified with the following form:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{response ~ L(glsw) + fixed + (random | group)
}\if{html}{\out{</div>}}

For more information, please see the \code{formula} subsection in details.}

\item{data}{The data.}

\item{bw}{A numeric value. It is the value of bandwidth or \code{"CV"}.
In this stage this function only support adaptive bandwidth.
And its unit must be the number of nearest neighbours.
If \code{"CV"} is specified, the algorithm will automatically select an
optimized bandwidth value.}

\item{kernel}{A character value. It specify which kernel function is used
in GWR part. Possible values are
\describe{
\item{\code{gaussian}}{Gaussian kernel function \eqn{k(d)=\exp\left(-\frac{d^2}{b^2}\right)}}
\item{\code{bisquared}}{Bi-squared kernel function. If \eqn{d<b} then \eqn{k(d)=\left(1-\frac{d^2}{b^2}\right)^2} else \eqn{k(d)=0}}
}}

\item{eps_iter}{A numeric value. Terminate threshold of back-fitting.}

\item{max_iters}{An integer value. The maximum of iteration.}

\item{max_retries}{An integer value. If the algorithm tends to be diverge,
it stops automatically after trying \emph{max_retires} times.}

\item{f_test}{A logical value. Determine whether to do F test on GLSW effects.
If \code{f_test=TURE}, there will be a \code{f_test} item in the returned object
showing the F test for each GLSW effect.}

\item{coords}{A 2-column matrix.
It consists of coordinates for each group.}

\item{alpha}{A numeric value. It is the size of the first trial step in
maximum likelihood algorithm.}

\item{eps_gradient}{A numeric value. Terminate threshold of
maximum likelihood algorithm.}

\item{ml_type}{An integer value. Represent which maximum likelihood
algorithm is used. Possible values are:
\describe{
\item{\code{D_Only}}{Only \eqn{D} is specified by maximum likelihood.}
\item{\code{D_Beta}}{Both \eqn{D} and \eqn{beta} is specified by maximum likelihood.}
}}
}
\value{
A list describing the model with following fields.
\describe{
\item{\code{gamma}}{Coefficients of group-level spatially weighted effects.}
\item{\code{beta}}{Coefficients of fixed effects.}
\item{\code{mu}}{Coefficients of sample-level random effects.}
\item{\code{D}}{Variance-covariance matrix of sample-level random effects.}
\item{\code{sigma}}{Variance of errors.}
\item{\code{effects}}{A list including names of all effects.}
\item{\code{call}}{Calling of this function.}
\item{\code{frame}}{The DataFrame object sent to this call.}
\item{\code{frame.parsed}}{Variables extracted from the data.}
\item{\code{groups}}{Unique group labels extracted from the data.}
\item{\code{f_test}}{A list of F test for GLSW effects. Only exists when \code{f_test=TRUE}. Each item contains the F value, degrees of freedom in the numerator, degrees of freedom in the denominator, and \eqn{p} value of \eqn{F>F_\alpha}.}
}
}
\description{
A Hierarchical Linear Model (HLM) with group-level geographically weighted effects.
}
\details{
\subsection{Effect Specification in Formula}{

In the HGWR model, there are three types of effects specified by the
\code{formula} argument:
\describe{
\item{Group-level spatially weighted (GLSW, aka. local fixed) effects}{Effects wrapped by functional symbol \code{L}.}
\item{Sample-level random (SLR) effects}{Effects specified outside the functional symbol \code{L} but to the left of symbol \code{|}.}
\item{Fixed effects}{Other effects}
}
For example, the following formula in the example of this function below is written as

\if{html}{\out{<div class="sourceCode r">}}\preformatted{y ~ L(g1 + g2) + x1 + (z1 | group)
}\if{html}{\out{</div>}}

where \code{g1} and \code{g2} are GLSW effects,
\code{x1} is the fixed effects,
and \code{z1} is the SLR effects grouped by the group indicator \code{group}.
Note that SLR effects can only be specified once!
}
}
\section{Functions}{
\itemize{
\item \code{spatial_hetero_test(hgwrm)}: Test the spatial heterogeneity with bootstrapping.

\item \code{hgwr_fit()}: Fit a HGWR model

}}
\examples{
data(mulsam.test)
hgwr(
  formula = y ~ L(g1 + g2) + x1 + (z1 | group),
  data = mulsam.test$data,
  coords = mulsam.test$coords,
  bw = 10
)

mod_Ftest <- hgwr(
  formula = y ~ L(g1 + g2) + x1 + (z1 | group),
  data = mulsam.test$data,
  coords = mulsam.test$coords,
  bw = 10
)
summary(mod_Ftest)

}
