% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgwr.R
\name{logLik.hgwrm}
\alias{logLik.hgwrm}
\title{Log likelihood function}
\usage{
\method{logLik}{hgwrm}(object, ...)
}
\arguments{
\item{object}{An \code{hgwrm} object.}

\item{\dots}{Additional arguments.}
}
\value{
An \code{logLik} instance used for S3 method \code{logLik()}.
}
\description{
Log likelihood function
}
