% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgwr.R
\name{matrix2char}
\alias{matrix2char}
\title{Convert a numeric matrix to character matrix according to a format string.}
\usage{
matrix2char(m, fmt = "\%.6f")
}
\arguments{
\item{m}{A numeric matrix.}

\item{fmt}{Format string. Passing to \code{\link[base:sprintf]{base::sprintf()}}.}
}
\description{
Convert a numeric matrix to character matrix according to a format string.
}
\seealso{
\code{\link[base:sprintf]{base::sprintf()}}, \code{\link[=print.hgwrm]{print.hgwrm()}}, \code{\link[=print.summary.hgwrm]{print.summary.hgwrm()}}.
}
