% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{mcmc_options}
\alias{mcmc_options}
\title{Markov Chain Monte Carlo Esitmator Options}
\usage{
mcmc_options(mcmc_iters = 5000, mcmc_burnin = 400)
}
\description{
Markov Chain Monte Carlo Esitmator Options
}
