% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.formula.R
\name{parse.formula}
\alias{parse.formula}
\alias{stack.push}
\alias{stack.pop}
\title{Parse a HGWR formula.}
\usage{
parse.formula(formula)

\method{stack}{push}(s, x)

\method{stack}{pop}(s)
}
\arguments{
\item{formula}{A \code{formula} object.
Its format is much like the formula used in \code{\link[lme4]{lmer}}
from package ''\strong{lme4}''.}

\item{s}{A \code{list}, \code{vector} or any other object which works with
function \code{\link[base]{c}}}

\item{x}{An object which can be appended to \code{s}.}
}
\value{
A \code{list} consists of:
- \code{response}: name of dependent (response) variable.
- \code{group}: name of group variable.
- \code{random.effects}: a vector of names of random effects.
- \code{fixed.effects}: a vector of names of fixed effects.
}
\description{
This function accept an R formula object and extract names of
the group variable, local fixed effects, global fixed effects
and random effects.
}
