% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hetero.R
\name{print.spahetbootres}
\alias{print.spahetbootres}
\title{Print the result of spatial heterogeneity test}
\usage{
\method{print}{spahetbootres}(x, ...)
}
\arguments{
\item{x}{A \code{spahetbootres} object.}

\item{\dots}{Other unused arguments.}
}
\description{
Print the result of spatial heterogeneity test
}
