% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgwr.R
\name{print.summary.hgwrm}
\alias{print.summary.hgwrm}
\title{Print summary of an \code{hgwrm} object.}
\usage{
\method{print}{summary.hgwrm}(x, decimal.fmt = "\%.6f", ...)
}
\arguments{
\item{x}{An object returned from \code{\link[=summary.hgwrm]{summary.hgwrm()}}.}

\item{decimal.fmt}{The format string passing to \code{\link[base:sprintf]{base::sprintf()}}.}

\item{...}{
  Arguments passed on to \code{\link[=print.table.md]{print.table.md}}
  \describe{
    \item{\code{col.sep}}{Column separator. Default to \code{""}.}
    \item{\code{header.sep}}{Header separator. Default to \code{"-"}.
If \code{header.sep} only contains one character, it will be repeated for each column.
If it contains more than one character, it will be printed below the first row.}
    \item{\code{row.begin}}{Character at the beginning of each row.
Default to \code{col.sep}.}
    \item{\code{row.end}}{Character at the ending of each row.
Default to \code{col.sep}.}
    \item{\code{table.before}}{Characters to be printed before the table.}
    \item{\code{table.after}}{Characters to be printed after the table.}
    \item{\code{table.style}}{Name of pre-defined style.
Possible values are \code{"plain"}, \code{"md"}, \code{"latex"}, or \code{"booktabs"}. Default to \code{"plain"}.}
  }}
}
\value{
No return.
}
\description{
Print summary of an \code{hgwrm} object.
}
\examples{
data(multisampling)
model <- hgwr(formula = y ~ L(g1 + g2) + x1 + (z1 | group),
              data = multisampling$data,
              coords = multisampling$coords,
              bw = 10)
summary(model)

}
\seealso{
\code{\link[=summary.hgwrm]{summary.hgwrm()}}, \code{\link[=print.table.md]{print.table.md()}}.
}
