% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgwr.R
\name{residuals.hgwrm}
\alias{residuals.hgwrm}
\title{Get residuals.}
\usage{
\method{residuals}{hgwrm}(object, ...)
}
\arguments{
\item{object}{An \code{hgwrm} object returned by \code{\link[=hgwr]{hgwr()}}.}

\item{...}{Parameter received from other functions.}
}
\value{
A vector consists of residuals.
}
\description{
Get residuals.
}
\seealso{
\code{\link[=hgwr]{hgwr()}}, \code{\link[=summary.hgwrm]{summary.hgwrm()}}, \code{\link[=coef.hgwrm]{coef.hgwrm()}} and \code{\link[=fitted.hgwrm]{fitted.hgwrm()}}.
}
