% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hetero.R
\name{spatial.hetero}
\alias{spatial.hetero}
\alias{spatial.hetero.matrix}
\alias{spatial.hetero.sf}
\alias{spatial.hetero.default}
\title{Generic method to test spatial heterogeneity}
\usage{
spatial.hetero(x, ...)

\method{spatial.hetero}{matrix}(x, coords, ...)

\method{spatial.hetero}{sf}(x, ...)

\method{spatial.hetero}{default}(x, ...)
}
\arguments{
\item{x}{The data to be tested.}

\item{...}{Additional arguments.}

\item{coords}{The coordinates used for testing.}
}
\description{
Generic method to test spatial heterogeneity
}
\section{Methods (by class)}{
\itemize{
\item \code{spatial.hetero(matrix)}: For the matrix, \code{coords} is necessary.

\item \code{spatial.hetero(sf)}: For the matrix, \code{coords} is necessary.

\item \code{spatial.hetero(default)}: Default behavior.

}}
