% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hetero.R
\name{spatial_hetero}
\alias{spatial_hetero}
\alias{spatial_hetero.matrix}
\alias{spatial_hetero.sf}
\alias{spatial_hetero.default}
\title{Generic method to test spatial heterogeneity}
\usage{
spatial_hetero(x, ...)

\method{spatial_hetero}{matrix}(x, coords, ...)

\method{spatial_hetero}{sf}(x, ...)

\method{spatial_hetero}{default}(x, ...)
}
\arguments{
\item{x}{The data to be tested.}

\item{...}{Additional arguments.}

\item{coords}{The coordinates used for testing.}
}
\description{
Generic method to test spatial heterogeneity
}
\section{Methods (by class)}{
\itemize{
\item \code{spatial_hetero(matrix)}: For the matrix, \code{coords} is necessary.

\item \code{spatial_hetero(sf)}: For the matrix, \code{coords} is necessary.

\item \code{spatial_hetero(default)}: Default behavior.

}}
