% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgwr.R
\name{summary.hgwrm}
\alias{summary.hgwrm}
\title{Summary an \code{hgwrm} object.}
\usage{
\method{summary}{hgwrm}(object, ...)
}
\arguments{
\item{object}{An \code{hgwrm} object returned from \code{\link[=hgwr]{hgwr()}}.}

\item{\dots}{Other arguments passed from other functions.}
}
\value{
A list containing summary informations of this \code{hgwrm} object
with the following fields.
\describe{
\item{\code{diagnostic}}{A list of diagnostic information.}
\item{\code{random.stddev}}{The standard deviation of random effects.}
\item{\code{random.corr}}{The correlation matrix of random effects.}
\item{\code{residuals}}{The residual vector.}
}
}
\description{
Summary an \code{hgwrm} object.
}
\seealso{
\code{\link[=hgwr]{hgwr()}}.
}
