\name{hgwrr-package}
\alias{hgwrr-package}
\alias{hgwrr}
\docType{package}
\title{
HGWR: Hierarchical and Geographically Weighted Regression
}
\description{
In this package we introduce an R and C++ implementation of Hierarchical 
and Geographically Weighted Regression (HGWR) model. This model divides
coefficients into three types: local fixed effects, global fixed effects, 
and random effects. If data have spatial hierarchical structures 
(especially are overlapping on some locations),
it is worth trying this model to reach better fitness.
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{hgwrr}
\packageIndices{hgwrr}
}
\author{
\packageAuthor{hgwrr}

Maintainer: \packageMaintainer{hgwrr}
}
\references{
Hu, Y., Lu, B., Ge, Y., Dong, G., 2022. 
Uncovering spatial heterogeneity in real estate prices via 
combined hierarchical linear model and geographically weighted regression. 
Environment and Planning B: Urban Analytics and City Science 239980832110638. 
\doi{10.1177/23998083211063885}
}
\keyword{ package }

