\name{hgwrr-package}
\alias{hgwrr-package}
\alias{hgwrr}
\docType{package}
\title{
\packageTitle{hgwrr}
}
\description{
\packageDescription{hgwrr}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{hgwrr}
\packageIndices{hgwrr}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{hgwrr}

Maintainer: \packageMaintainer{hgwrr}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
