% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hetero.R
\name{print.shgt}
\alias{print.shgt}
\title{Print the result of spatial heterogeneity test}
\usage{
\method{print}{shgt}(x, ...)
}
\arguments{
\item{x}{A \code{shgt} object.}

\item{\dots}{Other unused arguments.}
}
\description{
Print the result of spatial heterogeneity test
}
