% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hetero.R
\name{spatial_hetero_test}
\alias{spatial_hetero_test}
\title{Spatial heterogeneity test}
\usage{
spatial_hetero_test(
  x,
  coords,
  ...,
  resample = 5000,
  poly = 2,
  bw = 10,
  verbose = 0
)
}
\description{
Spatial heterogeneity test
}
